requirements.txt
paho-mqtt
azure-iot-device
requests
urllib